added a plugin for toggling the legend behaviour
In chart.js, clicking on the legend hides this legend. However, the average user is used to the oppositive behaviour. Using this code (grabbed from stackoverflow), we can focus on a particular dataset when clicking on the corresponding legend element.

Please note that it was only tested with line charts!